Conflicts:
	App/Config/routes.ini
	App/Config/settings.ini
	App/Controller/UserController.php
	App/Controller/WidgetController.php
	App/Entity/UserEntity.php
	System/CoraPHP/Http/Router.php
	System/CoraPHP/Model/ActiveRecord.php
	System/CoraPHP/Model/Database.php
	System/CoraPHP/Mvc/Controller.php

+++ System/CoraPHP/Model/ActiveRecord.php
@@ -52,7 +52,12 @@ abstract class ActiveRecord {
         return $list;
     }
     
-    public function make($data){
+    /**
+     * 
+     * @param array $data
+     * @return static
+     */
+    public function make($data = array()){
         
         $me = new static($this->db);
         
@@ -100,6 +105,18 @@ abstract class ActiveRecord {
         return null;
     }
     
+    public function getBy($field, $value){
+        
+        $data = $this->db->selectOne($this->getTable(), "*", "{$field} = '{$value}'");
+        
+        if($data)
+        {
+            return $this->make($data);
+        }
+        
+        return null;
+    }
+    
     /**
      * Default implementation of <b>save</b> method,
      * you can override if you need
@@ -120,14 +137,11 @@ abstract class ActiveRecord {
         }
     }
     
-    /**
-     * Default implementation of <b>delete</b> method,
-     * you can override if you need
-     * 
-     * @return boolean
-     */
     public function delete(){
         
-        return $this->db->delete($this->getTable(), "id = {$this->id}");
+        if($this->id !== null)
+        {
+            $this->db->delete($this->getTable(), "id = {$this->id}");
+        }
     }
 }
